Moving functionality to PullRequestService
