+++ src/GitHub.InlineReviews/Services/PullRequestSessionService.cs
@@ -24,7 +24,6 @@ using ReactiveUI;
 using Serilog;
 using PullRequestReviewEvent = Octokit.PullRequestReviewEvent;
 using static Octokit.GraphQL.Variable;
-using System.Windows.Forms;
 
 // GraphQL DatabaseId field are marked as deprecated, but we need them for interop with REST.
 #pragma warning disable CS0618
